Add files via upload
work in progress of the basic chore class. We will build upon this, add functionality outside of just toggling completion status, like changing the name of the assignee, etc. 

Next will need home class template that include chores as attributes. Maybe another class for users? This is just a start.